"free slice" command now refuses to free special slices like that which is returned from "sprite layer"
git-svn-id: https://rpg.hamsterrepublic.com/source/wip@2481 7d344553-34f0-0310-a9b1-970ce8f1c3a2

+++ yetmore.bas
@@ -1853,7 +1853,13 @@ SELECT CASE AS CONST id
   scriptret = find_plotslice_handle(SliceTable.ScriptSprite)
  CASE 361 '--free slice
   IF valid_plotslice(retvals(0), "free slice") THEN
-   DeleteSlice(@plotslices(retvals(0)))
+   DIM sl AS Slice Ptr
+   sl = plotslices(retvals(0))
+   IF sl->SliceType = slRoot OR sl->SliceType = slSpecial THEN
+    debug "free slice: cannot free " & SliceTypeName(sl) & " slice " & retvals(0)
+   ELSE
+    DeleteSlice(@plotslices(retvals(0)))
+   END IF
   END IF
  CASE 362 '--first child
   IF valid_plotslice(retvals(0), "first child") THEN
